move review status to helpers
